refactor(goals): extract transitions and portfolio engine
